add dagger implementation
