+++ KompasGitHubApp/app/build.gradle
@@ -72,6 +72,9 @@ dependencies {
 
     implementation "androidx.lifecycle:lifecycle-livedata-ktx:2.3.1"
 
+    implementation 'com.google.dagger:dagger:2.37'
+    kapt 'com.google.dagger:dagger-compiler:2.37'
+
 
     implementation 'io.reactivex.rxjava3:rxandroid:3.0.0'
     // Because RxAndroid releases are few and far between, it is recommended you also
@@ -98,4 +101,5 @@ dependencies {
     androidTestImplementation('com.adevinta.android:barista:4.2.0') {
         exclude group: 'org.jetbrains.kotlin' // Only if you already use Kotlin in your project
     }
+
 }
