+++ KompasGitHubApp/app/src/main/java/vanrrtech/app/kompasgithubapp/app/SearchFunction/Model/Networking/GitUserService.kt
@@ -1,8 +1,10 @@
-package vanrrtech.app.kompasgithubapp.app.SearchFunction.Model
+package vanrrtech.app.kompasgithubapp.app.SearchFunction.Model.Networking
 
 import retrofit2.http.GET
 import retrofit2.http.Path
 import retrofit2.http.Query
+import vanrrtech.app.kompasgithubapp.app.SearchFunction.Model.SearchResult
+import vanrrtech.app.kompasgithubapp.app.SearchFunction.Model.UserItem
 import vanrrtech.app.kompasgithubapp.app.UserDetailFunction.Model.UserRepoDetails
 
 interface GitUserService {
