+++ KompasGitHubApp/app/src/main/java/vanrrtech/app/kompasgithubapp/app/SearchFunction/Model/Networking/NetworkingModule.kt
@@ -1,0 +1,13 @@
+package vanrrtech.app.kompasgithubapp.app.SearchFunction.Model.Networking
+
+import dagger.Module
+import dagger.Provides
+import vanrrtech.app.kompasgithubapp.app.SearchFunction.Model.GitUserRemoteRepository
+
+@Module
+class NetworkingModule() {
+
+    @Provides
+    fun getGitUserRemoteRepositoru() = GitUserRemoteRepository()
+
+}
