+++ KompasGitHubApp/app/src/main/java/vanrrtech/app/kompasgithubapp/app/SearchFunction/ViewModel/SearchFunctionViewModel.kt
@@ -1,19 +1,24 @@
 package vanrrtech.app.kompasgithubapp.app.SearchFunction.ViewModel
 
 
+import android.app.Application
 import android.util.Log
-import androidx.lifecycle.LiveData
+import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.asLiveData
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.runBlocking
+import vanrrtech.app.kompasgithubapp.app.AppScope.MyApplication
 import vanrrtech.app.kompasgithubapp.app.SearchFunction.Model.GitUserRemoteRepository
 import vanrrtech.app.kompasgithubapp.app.SearchFunction.Model.UserItem
 
-class SearchFunctionViewModel : ViewModel() {
+class SearchFunctionViewModel(mRepo : GitUserRemoteRepository, application: Application) : AndroidViewModel(
+    application
+) {
 
-    val userItem = GitUserRemoteRepository.getUserRemoteRepo().getRecentUsers().asLiveData()
+
+    val userItem = mRepo.getRecentUsers().asLiveData()
     var searchItem  = MutableLiveData<UserItem?>()
 
     fun searchUser(name : String){
