--- KompasGitHubApp/app/src/main/java/vanrrtech/app/kompasgithubapp/app/Utils/MyApplication.kt
@@ -1,18 +1,0 @@
-package vanrrtech.app.kompasgithubapp.app.Utils
-
-import android.app.Application
-import android.content.Context
-
-
-class MyApplication : Application() {
-    override fun onCreate() {
-        super.onCreate()
-        context = this.applicationContext
-    }
-
-    companion object {
-        private var context: Context? = null
-        val appContext: Context?
-            get() = context
-    }
-}
