+++ KompasGitHubApp/build.gradle
@@ -1,6 +1,9 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
-    ext.kotlin_version = "1.5.0"
+    ext{
+        kotlin_version = "1.5.0"
+        dagger_version = '2.27.1'
+    }
     repositories {
         google()
         mavenCentral()
